+++ contracts/oracles/HistoricalOracle.sol
@@ -10,6 +10,9 @@ abstract contract HistoricalOracle is IHistoricalOracle {
         uint16 end;
         uint16 size;
         uint16 maxSize;
+        uint16 flags; // Bit flags for future use
+        uint112 __reserved; // Reserved for future use
+        uint64 extra; // For user extensions
     }
 
     mapping(address => BufferMetadata) internal observationBufferMetadata;